+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.3] - 2024-05-20
+
+* Fixes for item deletion in admin
+* Order pickup before dropoff in unassigned tasks
+
 ## [2.4.3] - 2024-05-07
 
 * First version for incidents + ability to report incidents from dispatch